+++ tests/test_cose.py
@@ -14,14 +14,14 @@ import pytest
 from cbor2 import CBORTag
 
 import cwt
-from cwt import COSE, EncodeError, cose_key
+from cwt import COSE, EncodeError, Recipient, cose_key
 
 from .utils import key_path
 
 
 @pytest.fixture(scope="session", autouse=True)
 def ctx():
-    return COSE()
+    return COSE(options={"kid_auto_inclusion": False})
 
 
 class TestCOSE:
@@ -29,10 +29,163 @@ class TestCOSE:
     Tests for COSE.
     """
 
+    def test_cose_constructor_with_invalid_args(self):
+        with pytest.raises(ValueError) as err:
+            COSE(options={"kid_auto_inclusion": "xxx"})
+            pytest.fail("COSE should fail.")
+        assert "kid_auto_inclusion should be bool" in str(err.value)
+
+    def test_cose_sample_cose_wg_examples_hmac_01(self, ctx):
+        cwt_str = "D8618543A10105A054546869732069732074686520636F6E74656E742E58202BDCC89F058216B8A208DDC6D8B54AA91F48BD63484986565105C9AD5A6682F6818340A20125044A6F75722D73656372657440"
+        key = cose_key.from_jwk(
+            {
+                "kty": "oct",
+                "alg": "HS256",
+                "kid": "our-secret",
+                "use": "sig",
+                "k": "hJtXIZ2uSN5kbQfbtTNWbpdmhkV8FJG-Onbc6mxCcYg",
+            }
+        )
+        token = ctx.encode_and_mac(
+            {},
+            {},
+            b"This is the content.",
+            key=key,
+            recipients=[Recipient(unprotected={1: -6, 4: b"our-secret"})],
+        )
+        assert token == bytes.fromhex(cwt_str)
+        assert ctx.decode(token, key) == b"This is the content."
+
+    def test_cose_sample_cose_wg_examples_aes_ccm_01(self, ctx):
+        cwt_str = "D8608443A1010AA1054D89F52F65A1C580933B5261A72F581C6899DA0A132BD2D2B9B10915743EE1F7B92A46802388816C040275EE818340A20125044A6F75722D73656372657440"
+        key = cose_key.from_jwk(
+            {
+                "kty": "oct",
+                "alg": "AES-CCM-16-64-128",
+                "kid": "our-secret",
+                "use": "enc",
+                "k": "hJtXIZ2uSN5kbQfbtTNWbg",
+            }
+        )
+        token = ctx.encode_and_encrypt(
+            {1: 10},
+            {5: bytes.fromhex("89F52F65A1C580933B5261A72F")},
+            b"This is the content.",
+            key=key,
+            nonce=bytes.fromhex("89F52F65A1C580933B5261A72F"),
+            recipients=[Recipient(unprotected={1: -6, 4: b"our-secret"})],
+        )
+        assert token == bytes.fromhex(cwt_str)
+        assert ctx.decode(token, key) == b"This is the content."
+
+    def test_cose_sample_cose_wg_examples_aes_gcm_01(self, ctx):
+        cwt_str = "D8608443A10101A1054C02D1F7E6F26C43D4868D87CE582460973A94BB2898009EE52ECFD9AB1DD25867374B3581F2C80039826350B97AE2300E42FC818340A20125044A6F75722D73656372657440"
+        key = cose_key.from_jwk(
+            {
+                "kty": "oct",
+                "alg": "A128GCM",
+                "kid": "our-secret",
+                "use": "enc",
+                "k": "hJtXIZ2uSN5kbQfbtTNWbg",
+            }
+        )
+        token = ctx.encode_and_encrypt(
+            {1: 1},
+            {5: bytes.fromhex("02D1F7E6F26C43D4868D87CE")},
+            b"This is the content.",
+            key=key,
+            nonce=bytes.fromhex("02D1F7E6F26C43D4868D87CE"),
+            recipients=[Recipient(unprotected={1: -6, 4: b"our-secret"})],
+        )
+        assert token == bytes.fromhex(cwt_str)
+        assert ctx.decode(token, key) == b"This is the content."
+
+    def test_cose_sample_cose_wg_examples_chacha_poly_01(self, ctx):
+        # cwt_str = "D8608444A1011818A1054C26682306D4FB28CA01B43B8058245F2BD5381BBB04921A8477E55C0D850069674A05E683D416583AA0CEE0E2929CDF648094818340A2012504477365632D32353640"
+        key = cose_key.from_jwk(
+            {
+                "kty": "oct",
+                "alg": "ChaCha20/Poly1305",
+                "kid": "sec-256",
+                "use": "enc",
+                "k": "Dx4tPEtaaXiHlqW0w9Lh8B8uPUxbanmIl6a1xNPi8QA",
+            }
+        )
+        token = ctx.encode_and_encrypt(
+            {},
+            {},
+            b"This is the content.",
+            key=key,
+            nonce=bytes.fromhex("26682306D4FB28CA01B43B80"),
+            recipients=[Recipient(unprotected={1: -6, 4: b"sec-256"})],
+        )
+        # assert token == bytes.fromhex(cwt_str)
+        assert ctx.decode(token, key) == b"This is the content."
+
+    def test_cose_sample_cose_wg_examples_chacha_poly_enc_01(self, ctx):
+        # cwt_str = "D08344A1011818A1054C5C3A9950BD2852F66E6C8D4F58243E536D4992A21591575C55FA22981B31AE1C045946D0E41A8A1ABD12BC9525922F4EB618"
+        key = cose_key.from_jwk(
+            {
+                "kty": "oct",
+                "alg": "ChaCha20/Poly1305",
+                "kid": "sec-256",
+                "use": "enc",
+                "k": "Dx4tPEtaaXiHlqW0w9Lh8B8uPUxbanmIl6a1xNPi8QA",
+            }
+        )
+        token = ctx.encode_and_encrypt(
+            {},
+            {},
+            payload=b"This is the content.",
+            key=key,
+            nonce=bytes.fromhex("5C3A9950BD2852F66E6C8D4F"),
+        )
+        # assert token == bytes.fromhex(cwt_str)
+        assert ctx.decode(token, key) == b"This is the content."
+
+    def test_cose_encode_and_mac_with_recipient_has_unsupported_alg(self, ctx):
+        key = cose_key.from_symmetric_key(alg="HS256")
+        with pytest.raises(NotImplementedError) as err:
+            ctx.encode_and_mac(
+                {},
+                {},
+                b"This is the content.",
+                key=key,
+                recipients=[Recipient(unprotected={1: 0, 4: b"our-secret"})],
+            )
+            pytest.fail("encode_and_mac should fail.")
+        assert "Algorithms other than direct are not supported for recipients." in str(
+            err.value
+        )
+
+    def test_cose_encode_and_encrypt_with_recipient_has_unsupported_alg(self, ctx):
+        key = cose_key.from_jwk(
+            {
+                "kty": "oct",
+                "alg": "AES-CCM-16-64-128",
+                "kid": "our-secret",
+                "use": "enc",
+                "k": "hJtXIZ2uSN5kbQfbtTNWbg",
+            }
+        )
+        with pytest.raises(NotImplementedError) as err:
+            ctx.encode_and_encrypt(
+                {1: 10},
+                {5: bytes.fromhex("89F52F65A1C580933B5261A72F")},
+                b"This is the content.",
+                key=key,
+                nonce=bytes.fromhex("89F52F65A1C580933B5261A72F"),
+                recipients=[Recipient(unprotected={1: 0, 4: b"our-secret"})],
+            )
+            pytest.fail("encode_and_encrypt should fail.")
+        assert "Algorithms other than direct are not supported for recipients." in str(
+            err.value
+        )
+
     def test_cose_encode_and_mac_with_invalid_payload(self, ctx):
         key = cose_key.from_symmetric_key(alg="HS256")
         with pytest.raises(EncodeError) as err:
-            ctx.encode_and_mac(b"", {}, datetime.datetime.now(), key, [{}, {}])
+            ctx.encode_and_mac({}, {}, datetime.datetime.now(), key)
             pytest.fail("encode_and_mac should fail.")
         assert "Failed to encode." in str(err.value)