+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changes
 Unreleased
 ----------
 
+- Add support for COSE algorithm names for KeyBuilder.from_jwk. `#72 <https://github.com/dajiaji/python-cwt/pull/72>`__
+- Add tests based on COSE WG examples. `#72 <https://github.com/dajiaji/python-cwt/pull/72>`__
+- Move parameter auto-gen function from CWT to COSE. `#72 <https://github.com/dajiaji/python-cwt/pull/72>`__
 - Refine COSE API to make the type of payload parameter be bytes only. `#71 <https://github.com/dajiaji/python-cwt/pull/71>`__
 - Simplify samples on docs. `#69 <https://github.com/dajiaji/python-cwt/pull/69>`__